enable instrumentation of de copies
We had a local patch to improve performance under some
workloads by reducing the frequency with which `de` objects
get copied. This patch introduces a build-tag-guarded
framework for enabling or disabling instrumentation of
how many de objects are being copied; if it's not present,
it's not used, but we still get at least an approximate
sense of the performance of trivial Set operations for
some workloads. If the build tag is added, we get numbers.