move more than one thing when overflowing
When overflowing one d into another, if we move only one object, we
can end up performing a lot of extra copies; for instance, if we're
moving things left, every time we move one item to the left, we
end up copying all but one of the items in the current d -- which
we already know is full when we're doing this. If we then want to
insert another item, and copy to the left, we do it again.

This patch adopts a balanced solution of trying to move up to
half of the available items into the available space at a time.
The actual number of items copied is usually roughly the same,
but we do it significantly less frequently. As N gets large, and
especially with large values of kd, this can become quite
significant. The impact isn't huge when using interfaces,
but it becomes more noticeable with concrete types. Using
a `make generic` btree converted to explicitly use int keys
and values:

	Linear:
	copied 1624997920 de for N=50000000, 165 ns/op =>
	copied 549999648 de for N=100000000, 153 ns/op

	Random:
	copied 1148275965 de for N=20000000, 723 ns/op =>
	copied 761915320 de for N=20000000, 656 ns/op